+++ nvim/.config/nvim/lua/editor/snacks.lua
@@ -88,22 +88,6 @@ require('snacks').setup {
           sort_empty = true,
         },
       },
-      scratch = {
-        win = {
-          input = {
-            keys = {
-              ['<c-a>'] = { 'scratch_new', mode = { 'n', 'i' } },
-              ['<c-d>'] = { 'scratch_delete', mode = { 'n', 'i' } },
-            },
-          },
-          list = {
-            keys = {
-              ['<c-a>'] = 'scratch_new',
-              ['<c-d>'] = 'scratch_delete',
-            },
-          },
-        },
-      },
     },
   },
 }
@@ -344,11 +328,3 @@ Snacks.toggle.scroll():map '<leader>uS'
 Snacks.toggle.inlay_hints():map '<leader>uh'
 Snacks.toggle.zoom():map '<leader>uZ'
 Snacks.toggle.zen():map '<leader>uz'
-
--- Scratch buffers
-vim.keymap.set('n', '<leader>.', function()
-  Snacks.scratch()
-end, { desc = 'Toggle Scratch Buffer' })
-vim.keymap.set('n', '<leader>S', function()
-  Snacks.scratch.select()
-end, { desc = 'Select Scratch Buffer' })